+++ src/main/java/msg/team1/Hi/domain/member/controller/MemberController.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import msg.team1.Hi.domain.member.dto.request.LoginRequest;
 import msg.team1.Hi.domain.member.dto.request.SignUpRequest;
 import msg.team1.Hi.domain.member.dto.response.MemberResponse;
-import msg.team1.Hi.domain.member.service.MemberService;
+import msg.team1.Hi.domain.member.service.MemberServiceImpl;
 import msg.team1.Hi.global.security.auth.MemberDetails;
 import msg.team1.Hi.global.security.jwt.properties.JwtProvider;
 import msg.team1.Hi.global.security.jwt.properties.dto.response.TokenResponse;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/member")
 public class MemberController {
 
-    private final MemberService memberService;
+    private final MemberServiceImpl memberService;
     private final JwtProvider jwtProvider;
 
     @PostMapping(value = "/signup", produces = MediaType.APPLICATION_JSON_VALUE)
