+++ src/main/java/msg/team1/Hi/domain/member/service/MemberService.java
@@ -1,57 +1,4 @@
 package msg.team1.Hi.domain.member.service;
 
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import msg.team1.Hi.domain.member.dto.request.LoginRequest;
-import msg.team1.Hi.domain.member.dto.request.SignUpRequest;
-import msg.team1.Hi.domain.member.dto.response.MemberResponse;
-import msg.team1.Hi.domain.member.entity.Member;
-import msg.team1.Hi.domain.member.repository.MemberRepository;
-import msg.team1.Hi.global.exception.collection.BadRequestException;
-import msg.team1.Hi.global.role.Role;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class MemberService {
-
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Transactional
-    public MemberResponse login(LoginRequest loginRequest) {
-        Member member = memberRepository
-                .findByEmail(loginRequest.getEmail())
-                .orElseThrow(() -> new BadRequestException("아이디 혹은 비밀번호를 확인하세요."));
-
-        boolean matches = passwordEncoder.matches(loginRequest.getPassword(), member.getPassword());
-        if(!matches) {
-            throw new BadRequestException("비밀번호가 일치하지 않습니다.");
-        }
-
-        return MemberResponse.of(member);
-    }
-
-    @Transactional
-    public MemberResponse signUp(SignUpRequest signUpRequest) {
-        boolean isExist = memberRepository.existsByEmail(signUpRequest.getEmail());
-        if(isExist) {
-            throw new BadRequestException("이미 존재하는 이메일입니다.");
-        }
-        String encodedPassword = passwordEncoder.encode(signUpRequest.getPassword());
-
-        Member signUpMember = new Member(signUpRequest.getEmail(),
-                encodedPassword , signUpRequest.getName() ,
-                signUpRequest.getNumber(),
-                Role.from(signUpRequest.getRole()));
-
-        signUpMember = memberRepository.save(signUpMember);
-
-        return MemberResponse.of(signUpMember);
-    }
+public interface MemberService {
 }
